Flask/requirements.txt
flask
flask_restx
pymongo
python-dotenv
pymysql